change project structure and build process
I added a Cakefile with various tasks for build and test process.
Coffee-script is know a normal dependency. Everytime a user installs the
package we compile the Coffee-Script and c++ code. Index of the package
moved from `geos.js` to `lib/index.js`. Everything should be written in
coffee and placed under `src/`. `npm install` calls the cake task
`install` that compiles the binary and the coffee code.